1. Modified synced_bcf_reader to work with merges for gVCF files. Main change - add a field in reader called last_valid_record_idx. This allows vcfmerge to split intervals and add it to reader->buffer by manipulating reader->nbuffer and reader->last_valid_record_idx 2. DEBUG flag in Makefile
TODO: FIXME
1. Terrible hack: Added boolean flag called is_split_record in bcf1_t
structure to determine whether a BCF record was created because of a
split in a gVCF interval. Used in a terribly hacky manner in vcfmerge.
Should be removed, once I have figured out how to obtain REF from
location.
2. When INFO and FORMAT fields are erased using update_info with NULL
arguments, I shift the elements of the info and fmt arrays and decrement
the values of n_info and n_fmt. Likely, we should set NULL pointers in
the deleted indices and not change n_info/fmt. This will be
faster;however, any applications using this must check for NULL
pointers as a proxy for deletion. Right now, vcfmerge does not work with
NULL pointers method, but works with the shifting method.